Move StaticInfo to components/index
