+++ components/index/StaticInfo.vue
